[TRA-254] Move matrixHeader to /matrix dir
