+++ src/shared/components/matrix/matrixHeader/matrixHeader.component.tsx
@@ -1,4 +1,4 @@
-import matrixLogo from '../../../images/matrix-logo.png';
+import matrixLogo from '../../../../images/matrix-logo.png';
 import { Header, Photo } from './matrixHeader.styles';
 
 export const MatrixHeader = () => {
